+++ Backend/DevopsTools/Dev/scripts/helper/env_config.sh
@@ -22,5 +22,9 @@ export DATA_DIR="${DATA_DIR:-${ENV_DIR}/data}"
 export CERTS_DIR="${CERTS_DIR:-${ENV_DIR}/certs}"
 export KEYPAIR_DIR="${KEYPAIR_DIR:-${ENV_DIR}/keypair}"
 
+export API_GATEWAY_CONTAINER_NAME="api-gateway"
 export BACKEND_CONTAINER_NAME="backend"
 export BACKEND_CONTAINER_PORT=60001
+
+export K8S_CLUSTER_NAME="learn-dot-net"
+export K8S_NAMESPACE="learn-dot-net"