+++ Backend/DevopsTools/Dev/scripts/k8s-clean.sh
@@ -9,14 +9,11 @@ MODE="dev"
 export MODE
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
 ENV_DIR="$(cd "${SCRIPT_DIR}/.." >/dev/null 2>&1 && pwd)"
-BACKEND_DIR="$(cd "${SCRIPT_DIR}/../../.." >/dev/null 2>&1 && pwd)"
 HELPER_DIR="${SCRIPT_DIR}/helper"
 export ENV_DIR
 
 ENV_FILE="${HELPER_DIR}/env_config.sh"
 FUNCTIONS_FILE="${HELPER_DIR}/functions.sh"
-CERT_SCRIPT="${HELPER_DIR}/generate_certs.sh"
-KEYPAIR_SCRIPT="${HELPER_DIR}/generate_keypair.sh"
 
 # -------------------------------
 # Load Environment & Helpers
@@ -26,26 +23,11 @@ KEYPAIR_SCRIPT="${HELPER_DIR}/generate_keypair.sh"
 source "${ENV_FILE}"
 # shellcheck source=scripts/helper/functions.sh
 source "${FUNCTIONS_FILE}"
-# shellcheck source=scripts/helper/generate_certs.sh
-source "${CERT_SCRIPT}"
-# shellcheck source=scripts/helper/generate_keypair.sh
-source "${KEYPAIR_SCRIPT}"
 
 # -------------------------------
-# Generate Environment & Certificates
+# Delete Kubernetes Resources
 # -------------------------------
 
-create_env_file
-create_files_from_templates
-create_data_folders
+delete_k8s_resources
 
-generate_root_ca
-generate_cert_with_keystore_and_truststore "api-gateway" "api-gateway" "${BACKEND_HOSTNAME}"
-
-# -------------------------------
-# Docker Image Build
-# -------------------------------
-
-build_backend_image "${BACKEND_DIR}"
-  
-echo "Initialize completed successfully."
+echo "Kubernetes resources deleted successfully."